# workflow/openapi.yaml
openapi: 3.1.0
info:
  title: Brand Forge n8n API
  version: 0.1.0
  summary: n8n webhook API for decoding brands and generating design specs, landing HTML, and marketing assets.
servers:
  - url: http://localhost:5678/webhook/brand
    description: Default local n8n webhook base
  - url: "{n8n_base_url}"
    description: Custom n8n base URL (include the /webhook/brand suffix)
    variables:
      n8n_base_url:
        default: http://localhost:5678/webhook/brand
        description: Base URL for the n8n brand webhook routes
paths:
...
  /html:
    post:
      operationId: generateHtml
      summary: Generate an 8-section landing page HTML from a design.md spec.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [brand_run_id, design_md]
              properties:
                brand_run_id:
                  type: string
                design_md:
                  type: string
      responses:
        '200':
          description: Landing page HTML and outline
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HtmlResponse'
        '502':
          description: Upstream error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    HtmlResponse:
      type: object
      required: [html, outline, tokens_total]
      properties:
        html:
          type: string
        outline:
          type: object
          properties:
            sections:
              type: array
              items:
                type: object
                additionalProperties: true
          additionalProperties: true
        tokens_total:
          type: integer
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        hint:
          type: string
        upstream_status:
          type: integer
        url:
          type: string